+++ bot/utilities/roleChecker.js
@@ -1,5 +1,5 @@
 'use strict';
-const _ = require('lodash');
+
 function capitalize(string) {
     return string.charAt(0).toUpperCase() + string.slice(1).toLowerCase();
 }